First step of xplat support: make the build work
