+++ accessibility-sys/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "accessibility-sys"
-version = "0.1.3"
-authors = ["Mack Straight <mack@discordapp.com>"]
+version = "0.2.0"
+authors = ["Mack Straight <mack@str.rs>"]
 edition = "2018"
 license = "MIT / Apache-2.0"
 homepage = "https://github.com/eiz/accessibility"
