+++ accessibility-sys/src/lib.rs
@@ -1,3 +1,5 @@
+#![cfg(target_os = "macos")]
+
 mod action_constants;
 mod attribute_constants;
 mod error;
