+++ accessibility/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "accessibility"
-version = "0.1.6"
-authors = ["Mack Straight <mack@discordapp.com>"]
+version = "0.2.0"
+authors = ["Mack Straight <mack@str.rs>"]
 edition = "2018"
 license = "MIT / Apache-2.0"
 homepage = "https://github.com/eiz/accessibility"
@@ -9,12 +9,13 @@ repository = "https://github.com/eiz/accessibility"
 description = "Bindings for macOS Accessibility services"
 
 [dependencies]
+
+[target.'cfg(target_os = "macos")'.dependencies]
 cocoa = "0.24"
 core-foundation = "0.9"
 objc = "0.2"
 thiserror = "1"
-
-accessibility-sys = { path = "../accessibility-sys", version = "0.1.3" }
+accessibility-sys = { path = "../accessibility-sys", version = "0.2.0" }
 
 [package.metadata.docs.rs]
 default-target = "x86_64-apple-darwin"
