+++ aq/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aq"
-version = "0.1.4"
+version = "0.2.0"
 authors = ["Mack Straight <mack@discordapp.com>"]
 edition = "2018"
 license = "MIT / Apache-2.0"
@@ -12,4 +12,4 @@ description = "Accessibility query"
 core-foundation = "0.9"
 structopt = "0.3"
 
-accessibility = { path = "../accessibility", version = "0.1.6" }
+accessibility = { path = "../accessibility", version = "0.2.0" }
